+++ CHANGELOG.md
@@ -14,6 +14,7 @@ All notable changes to `@homebridge/ciao` will be documented in this file. This 
 - fix: correct sign on local-compression pointer offset
 - refactor: drop redundant assertion in formatMappedIPv4Address
 - refactor: replace deprecated String#substr with slice
+- perf: use Date.now() instead of new Date().getTime()
 
 ## v1.3.8 (2026-05-04)