+++ CHANGELOG.md
@@ -2,6 +2,10 @@
 
 All notable changes to `@homebridge/ciao` will be documented in this file. This project tries to adhere to [Semantic Versioning](http://semver.org/).
 
+## v1.3.9 (Pending Release)
+
+- chore: dependency updates
+
 ## v1.3.8 (2026-05-04)
 
 - chore: dependency updates